New renderer with temporal dithering and interpolation.
The goal is to support larger matrices by shifting the balance of work
between the host and the firmware.  The host will be responsible for
whitepoint and gamma correction and send the resulting output to
the firmware at up to 11 bits per color component resolution (33 bits
per pixel).  This arrangement frees up memory previously used for
look-up tables that can now be used to support more pixels.

In addition, parameters of the matrix such as the number of led strips,
leds per strip, max dither bits, and rendering algorithm can all be
configured at runtime.  These parameters are currently hardcoded in
this first draft but later on they'll be transmitted by configuration
requests.

Cleaned up the config.h header a bit.

Hacked up the server to support the new protocol changes pending
deeper rework.