Rewrite LED driver timing loop to fix flickering.
The previous code from OctoWS2811 seems a bit flaky on some boards and the
timing for the first bit in the string can be misaligned. The new code
ensures that the DMA channel that sets the start bit is the first one
to be enabled and uses the T1H phase to update as few registers as
needed before the next cycle begins so it doesn't get cut short.

Updated built-in timings slightly for better stability.

Added a mechanism for generating oscilloscope trigger traces using
the TDO pin on the Fadecandy board.